Finished CMake refactor.  Only gcc done for now.
With this commit, compiling works, but the executable produced fails.
I'm not sure if it's because of the modifications I did to the build
system, but I don't think so.  I will continue testing and try to merge
the latest changes from the dev branch.

I will also adapt the *.cmake files for the other compilers