Use Cray compiler wrappers on xc50
This commit adjusts project/sles-15-skylake-64-xc50-intel.cmake
to use the compiler wrapers 'cc' and 'ftn' for all purposes
on the xc50 machines, not just the MPI compilation steps.

This fixes an error where compilation failed because the
fftw3 library could not be found, from an otherwise pristine
shell environment.  The Cray compiler wrappers understand
how to parse the loaded module, but the 'module load'
command does not make fftw visible to the unwrapped
compilers.