src/CMakeLists.txt: define CPLOCN_GLOB when configuring WITH_NEMO
In a following commit we will import the sources of the NEMO coupling
subroutines, so that GEM-NEMO (GEM with NEMO coupling) can be built
out-of-the-box from the MIG/gem repository.

To control whether we build with NEMO coupling or not, add a WITH_NEMO
cache variable, defaulting to false, and document it using 'set'.

If it is set, define a variable 'CPLOCN_GLOB' that globs the coupling
subroutine sources. This variable is used in the rpnphy CMakeLists to
conditionnally include these source files in the build of the rpnphy
target. Follow what was done for MACH by checking if the 'src/cplocn'
subdirectory is present when the option is passed, and add similar
status messages in c956fff4c (Issue #71 : initialise WITH_MACH properly
so that its cache is correctly set, 2024-12-24).