.gitlab-ci.yml: add job for packaging GEM-NEMO
Add a job to our CI pipeline to create the SSM package for GEM-NEMO,
following the build job added in the previous commit. Adjust the
top-level CMakeLists to name the package 'gem-nemo' when configured
'WITH_NEMO'. Note that we adjust both "NAME", so that "gem-nemo" is
substituted when processing control.json.in, as well as the argument to
'ec_package_name', which sets the name of the '.ssm' archive on disk.

Note that we must manually adjust PATH to point to the work directory of
the build:rhel-8-icelake-64-intel-2022.1.2-gem-nemo job, in order for
the program 'prgemnml' to be found. This is because .common_setup sets
GEM_WORK to "work-$GEM_ARCH" before adjusting PATH, but we want to use a
different name.