.gitlab-ci.yml: add deploy job for GEM-NEMO
The "staging" job invokes ci-stage-ssm.sh with arguments 'gem' and
'gem', which respectively correspond to the SSM domain in which to
deploy and the prefix of the packages to deploy. Since our GEM-NEMO
package job creates an SSM package named 'gem-nemo', this package is not
processed by the script. Add a second 'deploy' job for GEM-NEMO, and
adjust the arguments of ci-stage-ssm.sh accordingly.