scripts: runmod.sh, rungem.sh: allow running with Iris in coupled mode
Currently with the GOSSIP coupler, to run GEM coupled with NEMO
("GEM-NEMO"), it suffices to use an executable configured WITH_NEMO, and
to set cpl_ocn_L in the 'coupling' namelist. The model scripts are
coupling-agnostic. Both GEM and NEMO are launched from their own
invocation of 'mpirun' (from within 'r.run_in_parallel'), so that each
model runs in its own MPI_COMM_WORLD. The models communicate with the
coupler via TCP/IP, i.e. the coupling communications do not use MPI.

With the upcoming new coupler (Iris), GEM, NEMO and Iris must run in the
same MPI_COMM_WORLD, since Iris uses MPI to communicate with the models.
This implies that both models and Iris must be launched by the same
'mpirun' invocation. In order to do that while still keeping
development of the GEM and NEMO scripts in separate repositories, let's
use the new '-singleworld' mode of 'r.run_in_parallel' [1].

In this mode, several 'r.run_in_parallel' "slave" instances share a
common "instance directory" with a "master invocation". The master
invocation waits for all of its slaves to write their "slave resource
file" to the common instance directory, at which points the master
invocation re-invokes itself with all slave programs as the '-pgm'
argument, such that all executables from each slave are launched by the
same 'mpirun' invocation.

To support this scenario, add a '-cpl' argument to both 'runmod.sh' and
'rungem.sh', to indicate that we are running in coupled mode, and in
that case, invoke 'r.mpirun' (which is a link to 'r.run_in_parallel'
when running via the GEM Maestro module) with the necessary flags:
- '-instancedir' to pass the path of the shared "instance directory"
- '-instances' to pass the name of the GEM slave instance
- '-env CCARD_ARGS' so that the GEM executable gets the same value for
this variable in the self-invocation as when the slave is invoked

Default '-cpl' to 'no' in both runmod.sh and rungem.sh so that by
default we do not add additional r.run_in_parallel flags.

[1] RPN-SI/env-utils@316036e (Merge branch 'instances-single-world' into 'dev', 2025-08-01)