{,/src/gem/}CMakeLists: define variable to set GEM executable name
The name of the GEM executable (maingemdm) is hardcoded in the main
CMakeLists as well as in src/gem/CMakeLists.txt. In a following commit
we wish to tweak the executable name when compiling WITH_CPL, such that
both executables can be installed in the same SSM domain. To make this
change easier, use a 'gem_exe' variable, which we set in the toplevel
CMakeLists, to hold the name of the executable.