{,src/}CMakeLists.txt: append '-cpl' to GEM executable name when building WITH_CPL
In order for both the regular GEM executable as well as the GEM-NEMO
executable to be installed in the same SSM package, adjust the name of
the binary when building WITH_CPL by adding a '-cpl' suffix. As such,
adjust variable 'gem_exe' in src/CMakeLists when processing the
'WITH_CPL' option. This variable is used in src/gem/CMakeLists.txt to
set the executable name. Note that this variable is also defined and
used in the toplevel CMakeLists, so make sure to define it before adding
the 'src' subdirectory, and set it using PARENT_SCOPE in
src/CMakeLists.txt, ensuring the modified value is passed to the
toplevel.