CMakeLists: allow packaging GEM-NEMO along with GEM
In previous commits we added a WITH_CPL option to optionnally build
GEM-NEMO. For future GEM versions we want to package both the regular
build as well as the GEM-NEMO binary in the same SSM package. This is
possible using the CPack variable CPACK_INSTALL_CMAKE_PROJECTS [1],
which allows creating a package combining several CMake-built binary
trees.

Add a CPL_BUILD_DIR option to point to GEM-NEMO CMake build directory,
and if it is defined, set CPACK_INSTALL_CMAKE_PROJECTS before including
CPack, in order to install both the current build tree as well as the
one at ${CPL_BUILD_DIR}. This way, the SSM package creation is invoked
from the regular (non-coupled) GEM build directory, once GEM-NEMO is
also built.

The arguments to CPACK_INSTALL_CMAKE_PROJECTS are described at [2]: a
list of quadruplets (build directory, project name, install component,
destination directory). Here we use "ALL" for the 3rd argument,
indicating we want to package all install components of each project
(the "component" install option [3] is anyway not currently used in
GEM), and "/" for the 4th component, indicating we want both projects to
be installed at the root of the package.

We use "gem" and "gem-nemo" (via ${NAME}) as the 2nd argument to
distinguish both projects in the "cpack" output. This name does not have
to correspond to the actual CMake project name.

Note that any installed file from the first project that have the same
name in both projects are overwritten by the ones from the second
project. As such, install GEM-NEMO first, such that the packaged
*-config scripts are those from the regular GEM build.

[1] https://cmake.org/cmake/help/latest/guide/tutorial/Packaging%20Debug%20and%20Release.html
[2] https://cmake.org/cmake/help/latest/module/CPack.html#variable:CPACK_INSTALL_CMAKE_PROJECTS
[3] https://cmake.org/cmake/help/latest/command/install.html#common-options