.gitlab-ci.yml: name work directory following .common_setup for RHEL9
.common_setup sets GEM_WORK to work-$GEM_ARCH, where GEM_ARCH is
$OS-$COMPILER_SUITE-$COMPILER_VERSION. .eccc_setup_intel sets
COMPILER_SUITE=intel, and $OS is $ORDENV_PLAT in our environment.

However, the RHEL9 build job in our pipeline creates a work directory
named "work-rhel-9-graniterapids-64_inteloneapi-2025.1.0" (with
"inteloneapi" instead of "intel" and an underscore instead of a dash).
Fix that, such that the work symbolic link is named in the same way as
.common_setup sets GEM_WORK.