.gitlab-ci.yml: set environment variables necessary for cross-compiling on RHEL9
For RHEL9 jobs we compile on ppp6t, which has
ORDENV_PLAT=rhel-9-amd64-64, but compile for U3 machines, which have
ORDENV_PLAT=rhel-9-graniterapids-64. As mentioned in the previous
commit, ORDENV_PLAT is used to set the "$OS" component of GEM_ARCH, and
subsequently GEM_WORK. To get the correct value for these variables, set
ORDENV_PLAT prior to sourcing the Intel 2025 setup file, in the RHEL9
build and package jobs.

Also set EC_ARCH correctly, for correcteness, as otherwise it is set to
"Linux_x86-64/inteloneapi-2025.1.0" after sourcing the setup file, since
BASE_ARCH is set to Linux_x86-64 upon login.

Finally, set SSMUSE_PLATFORMS to include $ORDENV_PLAT, which makes sure
that packages published only for that platform, and not the generic
amd64 platforms, are correctly loaded. This will be important in a
subsequent commit when we add a job for building GEM-NEMO, since we need
the 'cnccpl' pacakge which is only published for
rhel-9-graniterapids-64.