Export GEM_SETUP_SCRIPT_NAME variable
This variable is then used in .initial_setup to use the proper .eccc_setup
file created in the GEM_WORK directory. This file is used in maestro
configuration when running GEM. As we now have several compiler suites and
versions, it needs to use the same setup file as the one used to compile.